feat: add sparkline charts, compact mode, error handling

+++ index.js
@@ -2,12 +2,13 @@
 
 const {version, description} = require('./package.json');
 const parseCliArgs = require('./src/parseCliArgs');
-const UI = require('./src/ui');
-const ChildProcess = require('./src/childProcess');
+const UI = require('./src/UI');
+const ChildProcess = require('./src/ChildProcess');
 
-parseCliArgs({version, description}, ({userCommand}) => {
+parseCliArgs({version, description}, ({userCommand, compact, interval}) => {
   const ui = new UI({
     command: userCommand.fullCommand,
+    compact,
     version
   });
 
@@ -15,20 +16,31 @@ parseCliArgs({version, description}, ({userCommand}) => {
 
   const processExit = (code = 0, message) => {
     ui.destroy();
-    userProcess.kill();
-
+    if (userProcess && !userProcess.killed) {
+      userProcess.kill();
+    }
     if (message) {
       console.log(message);
     }
-
     process.exit(code);
+  };
+
+  const printError = (err) => {
+    ui.addLog('');
+    //ui.addLog('ERROR:', {error: true});
+    ui.addLog(err, {error: true});
   };
 
   ui.on('exit', () => processExit(0));
   ui.on('scrolledUp', () => userProcess.enableOutputBuffer());
   ui.on('scrolledDown', () => userProcess.disableOutputBuffer());
 
-  userProcess.on('output', (output) => ui.addLog(output));
-  userProcess.on('exit', (message) => processExit(0, message));
-  userProcess.on('error', (err) => processExit(1, err));
+  userProcess.on('stdout', (output) => ui.addLog(output));
+  userProcess.on('stderr', (err) => printError(err));
+  userProcess.on('exit', (message) => printError(message));
+  userProcess.on('error', (err) => printError(err));
+
+  //test
+  setInterval(() => ui.addCpu((Math.sin(+new Date() / 1000) + 1) / 2 * 100), interval * 1000);
+  setInterval(() => ui.addMem(Math.random() * 100), interval * 1000);
 });
